@@ -1,87 +1,99 @@
 {
-  "name": "wmid",
-  "version": "0.1.0",
-  "description": "WMID stands for What Must I Do",
-  "author": "Dmitry ETS Klishev <Dmitry.Klishev@gmail.com>",
-  "homepage": "https://etsbox.ru",
-  "contributors": [
-    "Dmitry ETS Klishev <Dmitry.Klishev@gmail.com>"
-  ],
-  "repository": "https://github.com/ets777/wmid.git",
-  "private": true,
-  "scripts": {
-    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "front-start": "ng serve -o",
-    "front-build-prod": "ng build --configuration production",
-    "front-build-dev": "ng build --configuration development",
-    "front-build-dev-watch": "ng build --watch --configuration development",
-    "back-start-dev": "cross-env NODE_ENV=dev nest start",
-    "back-start-dev-watch": "cross-env NODE_ENV=dev nest start --watch",
-    "back-test-dev": "cross-env NODE_ENV=dev jest",
-    "back-build-prod": "cross-env NODE_ENV=prod nest build",
-    "back-start-prod": "cross-env NODE_ENV=prod node dist/backend/main",
-    "back-test-prod": "cross-env NODE_ENV=prod jest"
-  },
-  "dependencies": {
-    "@angular/animations": "^17.0.5",
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
-    "@angular/forms": "^17.0.5",
-    "@angular/material": "^16.2.12",
-    "@angular/material-moment-adapter": "^16.2.12",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
-    "@angular/router": "^17.0.5",
-    "@nestjs/common": "^9.0.0",
-    "@nestjs/config": "^2.3.2",
-    "@nestjs/core": "^9.0.0",
-    "@nestjs/jwt": "^10.0.3",
-    "@nestjs/passport": "^9.0.3",
-    "@nestjs/platform-express": "^9.0.0",
-    "@nestjs/sequelize": "^9.0.2",
-    "@nestjs/swagger": "^6.3.0",
-    "bcryptjs": "^2.4.3",
-    "bootstrap": "^5.2.3",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.0",
-    "cross-env": "^7.0.3",
-    "mysql2": "^3.3.3",
-    "ngx-cookie-service": "^17.0.0",
-    "passport-jwt": "^4.0.1",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.5.0",
-    "sequelize": "^6.31.1",
-    "tslib": "^2.3.0",
-    "zone.js": "^0.14.2"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.5",
-    "@angular/cli": "~17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
-    "@nestjs/cli": "^9.0.0",
-    "@nestjs/schematics": "^9.0.0",
-    "@nestjs/testing": "^9.4.2",
-    "@types/bcryptjs": "^2.4.2",
-    "@types/express": "^4.17.13",
-    "@types/jest": "29.5.1",
-    "@types/node": "18.16.12",
-    "@types/passport-jwt": "^3.0.8",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.0.1",
-    "jest": "29.5.0",
-    "jest-mock": "^29.5.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "source-map-support": "^0.5.20",
-    "supertest": "^6.1.3",
-    "ts-jest": "29.1.0",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "4.2.0",
-    "typescript": "^5.2.2"
-  }
+    "name": "wmid",
+    "version": "0.1.0",
+    "description": "WMID stands for What Must I Do",
+    "author": "Dmitry ETS Klishev <Dmitry.Klishev@gmail.com>",
+    "homepage": "https://etsbox.ru",
+    "contributors": [
+        "Dmitry ETS Klishev <Dmitry.Klishev@gmail.com>"
+    ],
+    "repository": "https://github.com/ets777/wmid.git",
+    "private": true,
+    "scripts": {
+        "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+        "front-start": "ng serve -o",
+        "clean-cache": "ng cache clean",
+        "front-build-prod": "ng build --configuration production",
+        "front-build-dev": "ng build --configuration development",
+        "front-build-dev-watch": "ng build --watch --configuration development",
+        "back-start-dev": "cross-env NODE_ENV=dev nest start",
+        "back-start-dev-watch": "cross-env NODE_ENV=dev nest start --watch",
+        "back-test-dev": "cross-env NODE_ENV=dev jest",
+        "back-build-prod": "cross-env NODE_ENV=prod nest build",
+        "back-start-prod": "cross-env NODE_ENV=prod node dist/backend/main",
+        "back-test-prod": "cross-env NODE_ENV=prod jest"
+    },
+    "dependencies": {
+        "@angular/animations": "^19.1.4",
+        "@angular/common": "^19.1.4",
+        "@angular/compiler": "^19.1.4",
+        "@angular/core": "^19.1.4",
+        "@angular/forms": "^19.1.4",
+        "@angular/platform-browser": "^19.1.4",
+        "@angular/platform-browser-dynamic": "^19.1.4",
+        "@angular/router": "^19.1.4",
+        "@capacitor/app": "7.0.0",
+        "@capacitor/cli": "7.0.1",
+        "@capacitor/core": "7.0.1",
+        "@capacitor/haptics": "7.0.0",
+        "@capacitor/keyboard": "7.0.0",
+        "@capacitor/status-bar": "7.0.0",
+        "@ionic/angular": "^8.4.1",
+        "@maskito/angular": "^3.2.1",
+        "@maskito/core": "^3.2.1",
+        "@maskito/kit": "^3.2.1",
+        "@nestjs/common": "^9.0.0",
+        "@nestjs/config": "^2.3.2",
+        "@nestjs/core": "^9.0.0",
+        "@nestjs/jwt": "^10.0.3",
+        "@nestjs/passport": "^9.0.3",
+        "@nestjs/platform-express": "^9.0.0",
+        "@nestjs/sequelize": "^9.0.2",
+        "@nestjs/swagger": "^6.3.0",
+        "bcryptjs": "^2.4.3",
+        "class-transformer": "^0.5.1",
+        "class-validator": "^0.14.0",
+        "cross-env": "^7.0.3",
+        "date-fns": "^4.1.0",
+        "date-fns-tz": "^3.2.0",
+        "mysql2": "^3.3.3",
+        "ngx-cookie-service": "^19.0.0",
+        "passport-jwt": "^4.0.1",
+        "reflect-metadata": "^0.1.13",
+        "rxjs": "^7.5.0",
+        "sequelize": "^6.31.1",
+        "tslib": "^2.3.0",
+        "zone.js": "^0.15.0"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^19.1.4",
+        "@angular/cli": "^19.1.4",
+        "@angular/compiler-cli": "^19.1.4",
+        "@ionic/angular-toolkit": "^12.1.1",
+        "@nestjs/cli": "^9.0.0",
+        "@nestjs/schematics": "^9.0.0",
+        "@nestjs/testing": "^9.4.2",
+        "@types/bcryptjs": "^2.4.2",
+        "@types/express": "^4.17.13",
+        "@types/jest": "29.5.1",
+        "@types/node": "18.16.12",
+        "@types/passport-jwt": "^3.0.8",
+        "@types/supertest": "^2.0.11",
+        "@typescript-eslint/eslint-plugin": "^8.21.0",
+        "@typescript-eslint/parser": "^8.21.0",
+        "eslint": "^9.18.0",
+        "eslint-plugin-no-relative-import-paths": "^1.5.5",
+        "jest": "29.5.0",
+        "jest-mock": "^29.5.0",
+        "karma": "^6.4.0",
+        "karma-chrome-launcher": "^3.1.0",
+        "karma-coverage": "^2.2.0",
+        "source-map-support": "^0.5.20",
+        "supertest": "^6.1.3",
+        "ts-jest": "29.1.0",
+        "ts-loader": "^9.2.3",
+        "ts-node": "^10.0.0",
+        "tsconfig-paths": "4.2.0",
+        "typescript": "^5.5.4"
+    }
 }
