@@ -1,97 +1,40 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { RouterModule } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { MatSelectModule } from '@angular/material/select';
-import { MatInputModule } from '@angular/material/input';
-import { MatButtonModule } from '@angular/material/button';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatMenuModule } from '@angular/material/menu';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { CookieService } from 'ngx-cookie-service';
-
 import { AppComponent } from './app.component';
-import { TopBarComponent } from './top-bar/top-bar.component';
-import { IndexPageComponent } from './index-page/index-page.component';
-import { TaskAddPageComponent } from './tasks/component/task-add-page/task-add-page.component';
-import { GrowthHistoryIndexBlockComponent } from './growth-history-index-block/growth-history-index-block.component';
-import { GrowthProgressIndexBlockComponent } from './growth-progress-index-block/growth-progress-index-block.component';
-import { SignInComponent } from './auth/component/sign-in/sign-in.component';
-import { GoalsIndexBlockComponent } from './goals-index-block/goals-index-block.component';
-import { MealsIndexBlockComponent } from './meals-index-block/meals-index-block.component';
-import { MealAddPageComponent } from './meal-add-page/meal-add-page.component';
-import { UserListComponent } from './users/user-list/user-list.component';
-
-import { AuthGuard } from './auth/guard/auth.guard';
-import { TokenInterceptorService } from './interceptors/token-interceptor.service';
+import { TokenInterceptorService } from './core/interceptors/token-interceptor.interceptor';
+import { IonicModule } from '@ionic/angular';
+import { CoreModule } from './core/core.module';
+import { AppRoutingModule } from './app-routing.module';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    ReactiveFormsModule,
-    MatAutocompleteModule,
-    MatSelectModule,
-    MatInputModule,
-    BrowserAnimationsModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatDatepickerModule,
-    MatSnackBarModule,
-    MatSlideToggleModule,
-    MatProgressBarModule,
-    MatProgressSpinnerModule,
-    MatTooltipModule,
-    MatMenuModule,
-    RouterModule.forRoot([
-      { path: '', component: IndexPageComponent, canActivate: [AuthGuard] },
-      {
-        path: 'add-task',
-        component: TaskAddPageComponent,
-        canActivate: [AuthGuard],
-      },
-      {
-        path: 'users/list',
-        component: UserListComponent,
-        canActivate: [AuthGuard],
-      },
-      {
-        path: 'add-meal',
-        component: MealAddPageComponent,
-        canActivate: [AuthGuard],
-      },
-      { path: 'sign-in', component: SignInComponent },
-    ]),
-  ],
-  declarations: [
-    AppComponent,
-    TopBarComponent,
-    IndexPageComponent,
-    TaskAddPageComponent,
-    GrowthHistoryIndexBlockComponent,
-    GrowthProgressIndexBlockComponent,
-    SignInComponent,
-    GoalsIndexBlockComponent,
-    MealsIndexBlockComponent,
-    MealAddPageComponent,
-    UserListComponent,
-  ],
-  bootstrap: [AppComponent],
-  providers: [
-    CookieService,
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: TokenInterceptorService,
-      multi: true,
-    },
-  ],
+    declarations: [
+        AppComponent,
+    ],
+    bootstrap: [
+        AppComponent,
+    ],
+    imports: [
+        BrowserModule,
+        ReactiveFormsModule,
+        BrowserAnimationsModule,
+        AppRoutingModule,
+        IonicModule.forRoot({}),
+        RouterModule.forRoot([]),
+        CoreModule,
+    ],
+    providers: [
+        CookieService,
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: TokenInterceptorService,
+            multi: true,
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+    ],
 })
-export class AppModule {}
+export class AppModule { }
