@@ -5,19 +5,19 @@ import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 
 async function bootstrap(): Promise<void> {
-  const app = await NestFactory.create(AppModule);
-  app.setGlobalPrefix('api');
-  app.enableCors();
+    const app = await NestFactory.create(AppModule);
+    app.setGlobalPrefix('api');
+    app.enableCors();
 
-  if (Number(process.env.SWAGGER) === 1) {
-    const config = new DocumentBuilder()
-      .setTitle('WMID-API')
-      .setVersion(process.env.npm_package_version)
-      .build();
-    const document = SwaggerModule.createDocument(app, config);
-    SwaggerModule.setup('/swagger', app, document);
-  }
+    if (Number(process.env.SWAGGER) === 1) {
+        const config = new DocumentBuilder()
+            .setTitle('WMID-API')
+            .setVersion(process.env.npm_package_version)
+            .build();
+        const document = SwaggerModule.createDocument(app, config);
+        SwaggerModule.setup('/swagger', app, document);
+    }
 
-  await app.listen(process.env.PORT);
+    await app.listen(process.env.PORT);
 }
 bootstrap();
