@@ -2,12 +2,15 @@ import { Module } from '@nestjs/common';
 import { RolesService } from './roles.service';
 import { RolesController } from './roles.controller';
 import { AuthModule } from '../auth/auth.module';
-import { DatabaseModule } from '@backend/database/database.module';
+import { SequelizeModule } from '@nestjs/sequelize';
+import { Role } from './roles.model';
+import { User } from '@backend/users/users.model';
+import { UserRole } from './user-roles.model';
 
 @Module({
-  providers: [RolesService],
-  controllers: [RolesController],
-  imports: [AuthModule, DatabaseModule],
-  exports: [RolesService],
+    providers: [RolesService],
+    controllers: [RolesController],
+    imports: [SequelizeModule.forFeature([Role, User, UserRole]), AuthModule],
+    exports: [RolesService],
 })
-export class RolesModule {}
+export class RolesModule { }
