@@ -1,50 +1,36 @@
-import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
-import { IRole } from './roles.interface';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateRoleDto } from './dto/create-role.dto';
-import { DB_CONNECTION } from '@backend/database/database.module';
-import { ResultSetHeader } from 'mysql2/promise';
+import { InjectModel } from '@nestjs/sequelize';
+import { Role } from './roles.model';
 
 @Injectable()
 export class RolesService {
-  constructor(@Inject(DB_CONNECTION) private mysqlConnection: any) {}
+  constructor(
+    @InjectModel(Role) private roleRepository: typeof Role,
+  ) { }
 
-  async createRole(dto: CreateRoleDto): Promise<IRole> {
-    const [result]: [ResultSetHeader] = await this.mysqlConnection.query(`
-      insert into usr_roles (code, name)
-      values ('${dto.code}', '${dto.name}')
-    `);
-
-    return {
-      ...dto,
-      id: result.insertId,
-    };
-  }
-
-  async deleteRole(code: string): Promise<number> {
-    const [result]: [ResultSetHeader] = await this.mysqlConnection.query(`
-      delete from usr_roles
-      where code = '${code.toUpperCase()}'
-    `);
-
-    return result.affectedRows;
-  }
-
-  async getRoleByCode(code: string): Promise<IRole> {
-    const [[role]]: [[IRole]] = await this.mysqlConnection.query(`
-      select id, code, name
-      from usr_roles
-      where code = '${code.toUpperCase()}'
-    `);
-
+  async createRole(dto: CreateRoleDto): Promise<Role> {
+    dto.code = dto.code.toUpperCase();
+    const role = await this.roleRepository.create(dto);
     return role;
   }
 
-  async getAllRoles(): Promise<IRole[]> {
-    const [roles]: [IRole[]] = await this.mysqlConnection.query(`
-      select id, code, name
-      from usr_roles
-    `);
+  async deleteRole(code: string): Promise<number> {
+    const affectedRows = await this.roleRepository.destroy({
+      where: { code: code.toUpperCase() },
+    });
+    return affectedRows;
+  }
 
+  async getRoleByCode(code: string): Promise<Role> {
+    const role = await this.roleRepository.findOne({
+      where: { code: code.toUpperCase() },
+    });
+    return role;
+  }
+
+  async getAllRoles(): Promise<Role[]> {
+    const roles = await this.roleRepository.findAll({ include: { all: true } });
     return roles;
   }
 
@@ -55,33 +41,14 @@ export class RolesService {
       throw new HttpException('Роль не найдена', HttpStatus.NOT_FOUND);
     }
 
-    const filteredData = {
-      name: data.name,
-    };
+    const [affectedRows] = await this.roleRepository.update(data, { where: { code } });
 
-    const updatedData = {
-      ...role,
-      ...filteredData,
-    };
-
-    const [result]: [ResultSetHeader] = await this.mysqlConnection.query(`
-      update usr_users
-      set name = '${updatedData.name}'
-      where code = '${code.toUpperCase()}'
-    `);
-
-    return result.affectedRows;
+    return affectedRows;
   }
 
-  async getUserRoles(username: string): Promise<IRole[]> {
-    const [roles]: [IRole[]] = await this.mysqlConnection.query(`
-      select r.code, r.name
-      from usr_roles r
-      join usr_userRoles ur on ur.roleId = r.id
-      join usr_users u on u.id = ur.userId
-      where u.username = '${username}'
-    `);
-
-    return roles;
-  }
+  // TODO: move it to user controller
+  // async getUserRoles(username: string): Promise<Role[]> {
+  //   const roles = await this.roleRepository.findAll({ where: { username } });
+  //   return roles;
+  // }
 }
