@@ -2,9 +2,9 @@ import { IsBoolean, IsNumber } from 'class-validator';
 import { CommonTaskAppointmentDto } from './common-task-appointment.dto';
 
 export class CreateTaskAppointmentDto extends CommonTaskAppointmentDto {
-  @IsNumber(null, { message: 'Должно быть числом' })
-  taskId: number;
+    @IsNumber(null, { message: 'Must be a number' })
+    taskPeriodId: number;
 
-  @IsBoolean({ message: 'Должно быть булевым значением' })
-  isAdditional?: boolean;
+    @IsBoolean({ message: 'Должно быть булевым значением' })
+    isAdditional?: boolean;
 }
