@@ -1,6 +1,10 @@
+/**
+ * TODO: rename this
+ */
+
 export enum Status {
-  APPOINTED = 1,
-  COMPLETED,
-  POSTPONED,
-  REJECTED,
+    APPOINTED = 1,
+    COMPLETED,
+    POSTPONED,
+    REJECTED,
 }
