@@ -15,19 +15,27 @@ import { CreateTaskCategoryDto } from './dto/create-task-category.dto';
 import { RolesGuard } from '@backend/auth/guards/roles.guard';
 import { AccessTokenGuard } from '@backend/auth/guards/accessToken.guard';
 import { UpdateTaskCategoryDto } from './dto/update-task-category.dto';
+import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { TaskCategory } from './task-categories.model';
 
+@ApiTags('Категории заданий')
 @Controller('task-categories')
 @UseGuards(AccessTokenGuard)
 export class TaskCategoriesController {
   constructor(private taskCategoriesService: TaskCategoriesService) {}
 
+  @ApiOperation({ summary: 'Создание категории задания' })
+  @ApiResponse({ status: 200, type: TaskCategory })
+  @ApiBody({ type: CreateTaskCategoryDto })
   @Roles('admin')
   @UseGuards(RolesGuard)
   @Post()
-  create(@Body() dto: CreateTaskCategoryDto): Promise<number> {
+  create(@Body() dto: CreateTaskCategoryDto): Promise<TaskCategory> {
     return this.taskCategoriesService.createTaskCategory(dto);
   }
 
+  @ApiOperation({ summary: 'Удаление категории задания' })
+  @ApiResponse({ status: 200, type: Number })
   @Roles('admin')
   @UseGuards(RolesGuard)
   @Delete('/:code')
@@ -35,16 +43,22 @@ export class TaskCategoriesController {
     return this.taskCategoriesService.deleteTaskCategory(code);
   }
 
+  @ApiOperation({ summary: 'Выбор категории задания по ID' })
+  @ApiResponse({ status: 200, type: TaskCategory })
   @Get('/:code')
-  getByCode(@Param('code') code: string): Promise<ITaskCategory> {
+  getByCode(@Param('code') code: string): Promise<TaskCategory> {
     return this.taskCategoriesService.getTaskCategoryByCode(code);
   }
 
+  @ApiOperation({ summary: 'Получение всех категорий заданий' })
+  @ApiResponse({ status: 200, type: [TaskCategory] })
   @Get()
-  getAll(): Promise<ITaskCategory[]> {
+  getAll(): Promise<TaskCategory[]> {
     return this.taskCategoriesService.getAllTaskCategories();
   }
 
+  @ApiOperation({ summary: 'Обновление категории заданий' })
+  @ApiResponse({ status: 200, type: Number })
   @Patch('/:code')
   update(
     @Param('code') code: string,
