@@ -1,28 +1,26 @@
 import {
-  CanActivate,
-  ExecutionContext,
-  HttpException,
-  HttpStatus,
-  Injectable,
+    CanActivate,
+    ExecutionContext,
+    HttpException,
+    HttpStatus,
+    Injectable,
 } from '@nestjs/common';
 import { TasksService } from '../tasks.service';
 
 @Injectable()
 export class AuthorGuard implements CanActivate {
-  constructor(private tasksService: TasksService) {}
+    constructor(private tasksService: TasksService) { }
 
-  async canActivate(context: ExecutionContext): Promise<boolean> {
-    try {
-      const req = context.switchToHttp().getRequest();
-      console.log(req.params);
-      console.log(req.user);
-      const taskId = req.params.id;
-      const userId = req.user.id;
-      const task = await this.tasksService.getTaskById(taskId);
+    async canActivate(context: ExecutionContext): Promise<boolean> {
+        try {
+            const req = context.switchToHttp().getRequest();
+            const taskId = req.params.id;
+            const userId = req.user.id;
+            const task = await this.tasksService.getTaskById(taskId);
 
-      return task.userId === userId;
-    } catch (e) {
-      throw new HttpException('Нет доступа', HttpStatus.FORBIDDEN);
+            return task.userId === userId;
+        } catch (e) {
+            throw new HttpException('Нет доступа', HttpStatus.FORBIDDEN);
+        }
     }
-  }
 }
