@@ -1,207 +1,207 @@
 export const appointNextTaskCases = [
-  {
-    time: '06:01:00',
-    date: '2000-01-01',
-    lastAppointmentId: 1,
-    expectedTaskId: 2,
-    description: '1.1. Назначение следующего задания',
-  },
-  {
-    time: '06:01:00',
-    date: '2000-01-01',
-    lastAppointmentId: 2,
-    expectedTaskId: 6,
-    description:
-      '1.2. Назначение следующего задания с промежуточными удалённым и неактивным',
-  },
-  {
-    time: '06:39:00',
-    date: '2000-01-01',
-    lastAppointmentId: 3,
-    expectedTaskId: 8,
-    description:
-      '1.3. Назначение следующего задания при наличии важного задания на время, когда времени достаточно',
-  },
-  {
-    time: '06:40:00',
-    date: '2000-01-01',
-    lastAppointmentId: 3,
-    expectedTaskId: undefined,
-    description:
-      '1.4. Назначение следующего задания при наличии важного задания на время, когда времени не достаточно, но выполнять задание на время ещё рано',
-  },
-  {
-    time: '06:51:00',
-    date: '2000-01-01',
-    lastAppointmentId: 3,
-    expectedTaskId: 9,
-    description:
-      '1.5. Назначение следующего задания при наличии важного задания на время, когда времени не достаточно, и можно начать выполнять задание на время',
-  },
-  {
-    time: '07:01:00',
-    date: '2000-01-01',
-    lastAppointmentId: 4,
-    expectedTaskId: 11,
-    description:
-      '1.6. Назначение следующего задания при разных периодичностях, задание может быть назначено',
-  },
-  {
-    time: '07:01:00',
-    date: '2000-01-03',
-    lastAppointmentId: 5,
-    expectedTaskId: undefined,
-    description:
-      '1.7. Назначение следующего задания при разных периодичностях, задание не может быть назначено',
-  },
-  {
-    time: '06:58:00',
-    date: '2000-01-02',
-    lastAppointmentId: 7,
-    expectedTaskId: 13,
-    description:
-      '1.8. Назначение следующего задания при наличии важного задания на время в цепочке и достаточном времени',
-  },
-  {
-    time: '06:59:00',
-    date: '2000-01-02',
-    lastAppointmentId: 7,
-    expectedTaskId: 14,
-    description:
-      '1.9. Назначение следующего задания при наличии важного задания на время в цепочке и недостаточном времени',
-  },
-  {
-    time: '07:01:00',
-    date: '2000-01-02',
-    lastAppointmentId: 7,
-    expectedTaskId: 14,
-    description:
-      '1.10. Назначение следующего задания при наличии важного задания на время в цепочке и просрочке',
-  },
-  {
-    time: '23:59:00',
-    date: '2000-01-02',
-    lastAppointmentId: 8,
-    expectedTaskId: 17,
-    description:
-      '1.11. Назначение следующего задания при наличии важного задания на время в цепочке до полуночи',
-  },
-  {
-    time: '00:01:00',
-    date: '2000-01-03',
-    lastAppointmentId: 8,
-    expectedTaskId: 17,
-    description:
-      '1.12. Назначение следующего задания при наличии важного задания на время в цепочке после полуночи',
-  },
-  {
-    time: '08:56:00',
-    date: '2000-01-02',
-    lastAppointmentId: 9,
-    expectedTaskId: 19,
-    description:
-      '1.13. Назначение следующего задания при выполнении задания на время раньше времени, но времени до него недостаточно',
-  },
-  {
-    time: '08:54:00',
-    date: '2000-01-02',
-    lastAppointmentId: 9,
-    expectedTaskId: 19,
-    description:
-      '1.14. Назначение следующего задания при выполнении задания на время раньше времени, времени до него достаточно',
-  },
-  {
-    time: '09:01:00',
-    date: '2000-01-03',
-    lastAppointmentId: 9,
-    expectedTaskId: 19,
-    description:
-      '1.15. Назначение следующего задания при выполнении задания на время с опозданием',
-  },
-  {
-    time: '06:58:00',
-    date: '2000-01-03',
-    lastAppointmentId: 10,
-    expectedTaskId: 21,
-    description:
-      '1.16. Назначение следующего задания при наличии неважного задания на время в цепочке и достаточном времени',
-  },
-  {
-    time: '06:59:00',
-    date: '2000-01-03',
-    lastAppointmentId: 10,
-    expectedTaskId: 21,
-    description:
-      '1.17. Назначение следующего задания при наличии неважного задания на время в цепочке и недостаточном времени',
-  },
-  {
-    time: '07:01:00',
-    date: '2000-01-03',
-    lastAppointmentId: 10,
-    expectedTaskId: 21,
-    description:
-      '1.18. Назначение следующего задания при наличии неважного задания на время в цепочке и просрочке',
-  },
-  {
-    time: '23:59:00',
-    date: '2000-01-03',
-    lastAppointmentId: 11,
-    expectedTaskId: 24,
-    description:
-      '1.19. Назначение следующего задания при наличии неважного задания на время в цепочке до полуночи',
-  },
-  {
-    time: '00:01:00',
-    date: '2000-01-04',
-    lastAppointmentId: 11,
-    expectedTaskId: 24,
-    description:
-      '1.20. Назначение следующего задания при наличии неважного задания на время в цепочке после полуночи',
-  },
-  {
-    time: '06:39:00',
-    date: '2000-01-04',
-    lastAppointmentId: 13,
-    expectedTaskId: 28,
-    description:
-      '1.21. Назначение следующего задания при наличии неважного задания на время, когда времени достаточно',
-  },
-  {
-    time: '06:40:00',
-    date: '2000-01-04',
-    lastAppointmentId: 13,
-    expectedTaskId: 28,
-    description:
-      '1.22. Назначение следующего задания при наличии неважного задания на время, когда времени не достаточно, но выполнять задание на время ещё рано',
-  },
-  {
-    time: '06:51:00',
-    date: '2000-01-04',
-    lastAppointmentId: 13,
-    expectedTaskId: 28,
-    description:
-      '1.23. Назначение следующего задания при наличии неважного задания на время, когда времени не достаточно, и можно начать выполнять задание на время',
-  },
-  {
-    time: '06:50:00',
-    date: '2000-01-05',
-    lastAppointmentId: 14,
-    expectedTaskId: 32,
-    description: '1.24. Назначение задания с наличием отложенного задания',
-  },
-  {
-    time: '05:01:00',
-    date: '2000-01-06',
-    lastAppointmentId: 14,
-    expectedTaskId: 31,
-    description: '1.25. Назначение отложенного задания',
-  },
-  {
-    time: '07:01:00',
-    date: '2000-01-06',
-    lastAppointmentId: 16,
-    expectedTaskId: 35,
-    description:
-      '1.26. Назначение следующего задания при наличии удалённого задания в цепочке',
-  },
+    {
+        time: '06:01:00',
+        date: '2000-01-01',
+        lastAppointmentId: 1,
+        expectedTaskId: 2,
+        description: '1.1. Назначение следующего задания',
+    },
+    {
+        time: '06:01:00',
+        date: '2000-01-01',
+        lastAppointmentId: 2,
+        expectedTaskId: 6,
+        description:
+            '1.2. Назначение следующего задания с промежуточными удалённым и неактивным',
+    },
+    {
+        time: '06:39:00',
+        date: '2000-01-01',
+        lastAppointmentId: 3,
+        expectedTaskId: 8,
+        description:
+            '1.3. Назначение следующего задания при наличии важного задания на время, когда времени достаточно',
+    },
+    {
+        time: '06:40:00',
+        date: '2000-01-01',
+        lastAppointmentId: 3,
+        expectedTaskId: undefined,
+        description:
+            '1.4. Назначение следующего задания при наличии важного задания на время, когда времени не достаточно, но выполнять задание на время ещё рано',
+    },
+    {
+        time: '06:51:00',
+        date: '2000-01-01',
+        lastAppointmentId: 3,
+        expectedTaskId: 9,
+        description:
+            '1.5. Назначение следующего задания при наличии важного задания на время, когда времени не достаточно, и можно начать выполнять задание на время',
+    },
+    {
+        time: '07:01:00',
+        date: '2000-01-01',
+        lastAppointmentId: 4,
+        expectedTaskId: 11,
+        description:
+            '1.6. Назначение следующего задания при разных периодичностях, задание может быть назначено',
+    },
+    {
+        time: '07:01:00',
+        date: '2000-01-03',
+        lastAppointmentId: 5,
+        expectedTaskId: undefined,
+        description:
+            '1.7. Назначение следующего задания при разных периодичностях, задание не может быть назначено',
+    },
+    {
+        time: '06:58:00',
+        date: '2000-01-02',
+        lastAppointmentId: 7,
+        expectedTaskId: 13,
+        description:
+            '1.8. Назначение следующего задания при наличии важного задания на время в цепочке и достаточном времени',
+    },
+    {
+        time: '06:59:00',
+        date: '2000-01-02',
+        lastAppointmentId: 7,
+        expectedTaskId: 14,
+        description:
+            '1.9. Назначение следующего задания при наличии важного задания на время в цепочке и недостаточном времени',
+    },
+    {
+        time: '07:01:00',
+        date: '2000-01-02',
+        lastAppointmentId: 7,
+        expectedTaskId: 14,
+        description:
+            '1.10. Назначение следующего задания при наличии важного задания на время в цепочке и просрочке',
+    },
+    {
+        time: '23:59:00',
+        date: '2000-01-02',
+        lastAppointmentId: 8,
+        expectedTaskId: 17,
+        description:
+            '1.11. Назначение следующего задания при наличии важного задания на время в цепочке до полуночи',
+    },
+    {
+        time: '00:01:00',
+        date: '2000-01-03',
+        lastAppointmentId: 8,
+        expectedTaskId: 17,
+        description:
+            '1.12. Назначение следующего задания при наличии важного задания на время в цепочке после полуночи',
+    },
+    {
+        time: '08:56:00',
+        date: '2000-01-02',
+        lastAppointmentId: 9,
+        expectedTaskId: 19,
+        description:
+            '1.13. Назначение следующего задания при выполнении задания на время раньше времени, но времени до него недостаточно',
+    },
+    {
+        time: '08:54:00',
+        date: '2000-01-02',
+        lastAppointmentId: 9,
+        expectedTaskId: 19,
+        description:
+            '1.14. Назначение следующего задания при выполнении задания на время раньше времени, времени до него достаточно',
+    },
+    {
+        time: '09:01:00',
+        date: '2000-01-03',
+        lastAppointmentId: 9,
+        expectedTaskId: 19,
+        description:
+            '1.15. Назначение следующего задания при выполнении задания на время с опозданием',
+    },
+    {
+        time: '06:58:00',
+        date: '2000-01-03',
+        lastAppointmentId: 10,
+        expectedTaskId: 21,
+        description:
+            '1.16. Назначение следующего задания при наличии неважного задания на время в цепочке и достаточном времени',
+    },
+    {
+        time: '06:59:00',
+        date: '2000-01-03',
+        lastAppointmentId: 10,
+        expectedTaskId: 21,
+        description:
+            '1.17. Назначение следующего задания при наличии неважного задания на время в цепочке и недостаточном времени',
+    },
+    {
+        time: '07:01:00',
+        date: '2000-01-03',
+        lastAppointmentId: 10,
+        expectedTaskId: 21,
+        description:
+            '1.18. Назначение следующего задания при наличии неважного задания на время в цепочке и просрочке',
+    },
+    {
+        time: '23:59:00',
+        date: '2000-01-03',
+        lastAppointmentId: 11,
+        expectedTaskId: 24,
+        description:
+            '1.19. Назначение следующего задания при наличии неважного задания на время в цепочке до полуночи',
+    },
+    {
+        time: '00:01:00',
+        date: '2000-01-04',
+        lastAppointmentId: 11,
+        expectedTaskId: 24,
+        description:
+            '1.20. Назначение следующего задания при наличии неважного задания на время в цепочке после полуночи',
+    },
+    {
+        time: '06:39:00',
+        date: '2000-01-04',
+        lastAppointmentId: 13,
+        expectedTaskId: 28,
+        description:
+            '1.21. Назначение следующего задания при наличии неважного задания на время, когда времени достаточно',
+    },
+    {
+        time: '06:40:00',
+        date: '2000-01-04',
+        lastAppointmentId: 13,
+        expectedTaskId: 28,
+        description:
+            '1.22. Назначение следующего задания при наличии неважного задания на время, когда времени не достаточно, но выполнять задание на время ещё рано',
+    },
+    {
+        time: '06:51:00',
+        date: '2000-01-04',
+        lastAppointmentId: 13,
+        expectedTaskId: 28,
+        description:
+            '1.23. Назначение следующего задания при наличии неважного задания на время, когда времени не достаточно, и можно начать выполнять задание на время',
+    },
+    {
+        time: '06:50:00',
+        date: '2000-01-05',
+        lastAppointmentId: 14,
+        expectedTaskId: 32,
+        description: '1.24. Назначение задания с наличием отложенного задания',
+    },
+    {
+        time: '05:01:00',
+        date: '2000-01-06',
+        lastAppointmentId: 14,
+        expectedTaskId: 31,
+        description: '1.25. Назначение отложенного задания',
+    },
+    {
+        time: '07:01:00',
+        date: '2000-01-06',
+        lastAppointmentId: 16,
+        expectedTaskId: 35,
+        description:
+            '1.26. Назначение следующего задания при наличии удалённого задания в цепочке',
+    },
 ];
